Excellent schema progress for the day.
Figured out how to deal with teh validator and the v4 draft. There are some "annoying" properties to go around. The issue arrises when an individual wants to make a relatively clear schema, but is forced to always add "required." For most WIN purposes, you wouldn't make a schema property you didn't want stored. In that spirit, there is a convenient method that formates a more sensible WIN schema into a more wordy but valid v4 JSON schema.

Most of this commit takes care of a lot of what you would think is "common sense" behavior. Needs extensive testing, but it's fairly robust in anecdotal testing. I won't believe it until it has official tests. Not enough time for something super extensive.