Big array update. Now with lots of array support. Changes reference objects as well (have a schema path and an object path).
The important thing is that arrays can now be used, and there is support for various types of array behavior (checked for $ref valid behavior too). Changed up some logic for moving/changing key information.

This will all need to be cleaned up, but it's in a pretty useable state right now.